requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mlx==0.31.1
mlx-metal==0.31.1
mlx-whisper==0.4.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
Pygments==2.19.2
json-repair==0.46.1
mcp==1.27.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
